requirements.txt
Django==4.2.0
dnspython==2.3.0
requests==2.31.0
python-dotenv==1.0.0
whois==0.9.7
beautifulsoup4==4.12.2
